Update patch set 5
Patch Set 5: Code-Review-1

Actually, GNU coreutils only has '-a' (equivelent to '-p' with reduced logging).  POSIX doesn't have '-a'.  Therefore, I think we should continue to print with '-p' (obey the POSIX standard this was originally written for) and not print with '-a' (obey the GNU option that doesn't exist in POSIX)

Patch-set: 5