Improve run-tests-on-android.sh for sh_test usage.
Show a summary at the end, and return a meaningful exit status.

This doesn't actually work yet because it needs an upstream fix to use
$(()) rather than $[] for shell arithmetic --- mksh only supports the
former (and bash calls the latter deprecate and threatens to remove it).

Also run each test in its own subdirectory. The cp and mkdir tests, for
example, currently tread on each others' toes, even when run
sequentially, because cp leaves around a `walrus` that mkdir doesn't
expect to exist.

Bug: http://b/135744229
Test: run this script
Change-Id: I5f28340c9ea1f9cebb35175801197715f10700a4