Build vi.
I'm not sure this is ready for a symlink yet, but it is usable, and it's
the best editor we have available, and I have a long-outstanding bug
requesting vi that the machines are nagging me about.

Notably missing last time I used this vi: ZZ (have to :wq) and R/r (have
to insert and delete separately). But I did manage to make the edit
without using adb pull/push and editing on the host, so...

Bug: http://b/110170945
Change-Id: I0d400e5a52b07f3161d1be4c414480e84108ee35